LRCI-4119 Move Java file searching/recording to JUnitBatchTestClassGroup
